Fixed a couple broken references to directories

+++ Map/Map/Interface/PreferencesPanel/AdvancedPrefs/SliderHelper.c
@@ -107,7 +107,7 @@ char *timeUnitTupleLongForm(struct timeUnitTuple tuple, char* longForm) {
 }
 
 char *getPluralChar(float number, char *unit) {
-    int length = strlen(unit);
+    int length = (int)strlen(unit);
     char *retUnit = malloc((length + 2) * sizeof(char));
     if ((int)number > 1) {
         sprintf(retUnit, "%ss", unit);
